refactor(templates): modularize remix-supabase-bookmarks message components
